cur.c: use according 'int' type for "*" format specifier
Type mismatch detected by gcc:

    cur.c:361:31: warning: field width specifier '*' expects argument of type 'int', but argument 6 has type 'size_t' {aka 'long unsigned int'} [-Wformat=]
      361 |     mvwprintw(cmdh, 0, 0, "%s%*s", "Command", COLS-strlen("Command"), ski_id);
          |                              ~^~              ~~~~~~~~~~~~~~~~~~~~~~
          |                               |                   |
          |                               int                 size_t {aka long unsigned int}

+++ src/cur.c
@@ -358,7 +358,7 @@ fflush(dbg);
     cmdwSetStatusCur("");
 #else
     wstandout(cmdh);
-    mvwprintw(cmdh, 0, 0, "%s%*s", "Command", COLS-strlen("Command"), ski_id);
+    mvwprintw(cmdh, 0, 0, "%s%*s", "Command", (int)(COLS-strlen("Command")), ski_id);
     wstandend(cmdh);
 #endif
 }
@@ -519,7 +519,7 @@ void cmdwSetStatusCur(const char *msg)
     wstandout(cmdh);
     (void)sprintf(buf, "Command %2d/%-2d  %s", viewPid+2, 5, msg);
     mvwaddstr(cmdh, 0, 0, buf);
-    mvwprintw(cmdh, 0, strlen(buf), "%*s", COLS-strlen(buf), ski_id);
+    mvwprintw(cmdh, 0, strlen(buf), "%*s", (int)(COLS-strlen(buf)), ski_id);
     wstandend(cmdh);
     wrefresh(cmdh);
 }
